Seperated censor
